Add logic to calculate direct walk route
