+++ shared/src/commonMain/kotlin/cl/emilym/sinatra/domain/navigation/CalculateJourneyUseCase.kt
@@ -1,33 +1,23 @@
-package cl.emilym.sinatra.domain
+package cl.emilym.sinatra.domain.navigation
 
 import cl.emilym.sinatra.RouterException
 import cl.emilym.sinatra.data.models.Cachable
 import cl.emilym.sinatra.data.models.Journey
 import cl.emilym.sinatra.data.models.JourneyLeg
-import cl.emilym.sinatra.data.models.MapLocation
-import cl.emilym.sinatra.data.models.RouteServiceAccessibility
-import cl.emilym.sinatra.data.models.ServiceBikesAllowed
 import cl.emilym.sinatra.data.models.ServiceId
-import cl.emilym.sinatra.data.models.ServiceWheelchairAccessible
 import cl.emilym.sinatra.data.models.Stop
 import cl.emilym.sinatra.data.models.StopWithDistance
-import cl.emilym.sinatra.data.models.Time
 import cl.emilym.sinatra.data.models.distance
 import cl.emilym.sinatra.data.models.map
 import cl.emilym.sinatra.data.models.startOfDay
 import cl.emilym.sinatra.data.repository.NetworkGraphRepository
-import cl.emilym.sinatra.data.repository.RouteRepository
 import cl.emilym.sinatra.data.repository.RoutingPreferencesRepository
 import cl.emilym.sinatra.data.repository.StopRepository
 import cl.emilym.sinatra.data.repository.TransportMetadataRepository
+import cl.emilym.sinatra.domain.ActiveServicesUseCase
 import cl.emilym.sinatra.e
-import cl.emilym.sinatra.router.ArrivalBasedRouter
-import cl.emilym.sinatra.router.DepartureBasedRouter
 import cl.emilym.sinatra.router.RaptorConfig
-import cl.emilym.sinatra.router.RaptorJourney
-import cl.emilym.sinatra.router.RaptorJourneyConnection
 import cl.emilym.sinatra.router.RaptorStop
-import cl.emilym.sinatra.router.Router
 import cl.emilym.sinatra.router.RouterPrefs
 import cl.emilym.sinatra.router.Seconds
 import cl.emilym.sinatra.router.data.NetworkGraph
@@ -44,22 +34,6 @@ import kotlin.time.Duration
 import kotlin.time.Duration.Companion.days
 import kotlin.time.Duration.Companion.seconds
 
-data class JourneyLocation(
-    val location: MapLocation,
-    val exact: Boolean
-)
-
-sealed interface JourneyCalculationTime {
-    val time: Instant
-
-    data class DepartureTime(
-        override val time: Instant
-    ): JourneyCalculationTime
-    data class ArrivalTime(
-        override val time: Instant
-    ): JourneyCalculationTime
-}
-
 @Factory
 class CalculateJourneyUseCase(
     private val networkGraphRepository: NetworkGraphRepository,
@@ -70,6 +44,7 @@ class CalculateJourneyUseCase(
     private val routingPreferencesRepository: RoutingPreferencesRepository,
     private val routerFactory: RouterFactory,
     private val reconstructJourneyUseCase: ReconstructJourneyUseCase,
+    private val directWalkingJourneyUseCase: DirectWalkingJourneyUseCase
 ) {
 
     private val lock = Mutex()
@@ -112,6 +87,15 @@ class CalculateJourneyUseCase(
 
                 val computationCutoffTime = clock.now() + config.item.maximumComputationTime
                 val options = mutableListOf<Journey>()
+
+                directWalkingJourneyUseCase(
+                    departureLocation,
+                    arrivalLocation,
+                    anchorTime,
+                    getGraph().item.metadata.assumedWalkingSecondsPerKilometer.toInt().seconds
+                )?.let {
+                    options += it
+                }
 
                 for (o in config.item.options) {
                     val maximumWalkingTime = minOf(
@@ -210,145 +194,3 @@ class CalculateJourneyUseCase(
 
 }
 
-@Factory
-class ReconstructJourneyUseCase(
-    private val routeRepository: RouteRepository,
-    private val transportMetadataRepository: TransportMetadataRepository,
-    private val stopRepository: StopRepository
-) {
-
-    suspend operator fun invoke(
-        raptorJourney: RaptorJourney,
-        departureLocation: JourneyLocation,
-        arrivalLocation: JourneyLocation,
-        anchorTime: JourneyCalculationTime,
-        graph: NetworkGraph
-    ): Journey {
-        val connections = raptorJourney.connections
-        val startOfDay = anchorTime.time.startOfDay(transportMetadataRepository.timeZone())
-        val stops = stopRepository.stops()
-
-        val routes = routeRepository.routes(
-            connections.filterIsInstance<RaptorJourneyConnection.Travel>().map { it.routeId }
-        )
-        var legs = mutableListOf<JourneyLeg>()
-        var lastEndTime: Duration? = when (val connection = connections.first()) {
-            is RaptorJourneyConnection.Travel -> connection.endTime.seconds
-            else -> null
-        }
-        var lastStartOfDay = startOfDay
-
-        for (i in connections.indices) {
-            val stops = connections[i].stops.mapNotNull { s -> stops.item.firstOrNull { it.id == s } }
-            legs += when (val connection = connections[i]) {
-                is RaptorJourneyConnection.Travel -> {
-                    lastStartOfDay = startOfDay + connection.dayIndex.days
-                    lastEndTime = connection.endTime.seconds
-                    JourneyLeg.Travel(
-                        stops,
-                        (connection.endTime - connection.startTime).seconds,
-                        routes.item.first { it?.id == connection.routeId }!!,
-                        connection.heading,
-                        Time.create(connection.startTime.seconds, lastStartOfDay),
-                        Time.create(connection.endTime.seconds, lastStartOfDay),
-                        RouteServiceAccessibility(
-                            if (connection.bikesAllowed) ServiceBikesAllowed.ALLOWED else ServiceBikesAllowed.DISALLOWED,
-                            if (connection.wheelchairAccessible) ServiceWheelchairAccessible.ACCESSIBLE else ServiceWheelchairAccessible.INACCESSIBLE,
-                        )
-                    )
-                }
-                is RaptorJourneyConnection.Transfer -> when {
-                    lastEndTime != null -> JourneyLeg.Transfer(
-                        stops,
-                        connection.travelTime.seconds,
-                        Time.create(lastEndTime, lastStartOfDay),
-                        Time.create(lastEndTime + connection.travelTime.seconds, lastStartOfDay)
-                    ).also {
-                        lastEndTime += connection.travelTime.seconds
-                    }
-                    else -> {
-                        val inbetweenTime = connections
-                            .drop(i)
-                            .takeWhile { it is RaptorJourneyConnection.Travel }
-                            .sumOf { it.travelTime }
-                        val concreteTime = ((connections
-                            .drop(i)
-                            .first { it is RaptorJourneyConnection.Travel } as RaptorJourneyConnection.Travel)
-                            .startTime - inbetweenTime).seconds
-                        lastStartOfDay = startOfDay + (connections
-                            .drop(i)
-                            .first { it is RaptorJourneyConnection.Travel } as RaptorJourneyConnection.Travel)
-                            .dayIndex.days
-
-                        JourneyLeg.Transfer(
-                            stops,
-                            connection.travelTime.seconds,
-                            Time.create(concreteTime - connection.travelTime.seconds, lastStartOfDay),
-                            Time.create(concreteTime, lastStartOfDay)
-                        )
-                    }
-                }
-            }
-        }
-
-        if (!departureLocation.exact) {
-            run {
-                legs = legs.dropWhile { it is JourneyLeg.Transfer }.toMutableList()
-                if (legs.isEmpty()) return@run
-                val attachedStop = (legs.first { it is JourneyLeg.RouteJourneyLeg } as JourneyLeg.RouteJourneyLeg).stops.first()
-                val time = distance(departureLocation.location, attachedStop.location) * graph.metadata.assumedWalkingSecondsPerKilometer.toLong()
-                legs.add(0, JourneyLeg.TransferPoint(
-                    time.seconds,
-                    departureTime = legs.first().departureTime - time.seconds,
-                    arrivalTime = legs.first().departureTime
-                ))
-            }
-        }
-
-        if (!arrivalLocation.exact) {
-            run {
-                legs = legs.dropLastWhile { it is JourneyLeg.Transfer }.toMutableList()
-                if (legs.isEmpty()) return@run
-                val lastLeg = legs.last { it is JourneyLeg.RouteJourneyLeg } as JourneyLeg.RouteJourneyLeg
-                val attachedStop = lastLeg.stops.last()
-                val time = distance(arrivalLocation.location, attachedStop.location) * graph.metadata.assumedWalkingSecondsPerKilometer.toLong()
-                legs.add(JourneyLeg.TransferPoint(
-                    time.seconds,
-                    departureTime = lastLeg.arrivalTime,
-                    arrivalTime = lastLeg.arrivalTime + time.seconds
-                ))
-            }
-        }
-
-        return Journey(legs)
-    }
-
-}
-
-@Factory
-class RouterFactory {
-
-    operator fun invoke(
-        anchorTime: JourneyCalculationTime,
-        graph: NetworkGraph,
-        services: List<Cachable<List<ServiceId>>>,
-        config: RaptorConfig,
-        prefs: RouterPrefs
-    ): Router {
-        return when (anchorTime) {
-            is JourneyCalculationTime.ArrivalTime -> ArrivalBasedRouter(
-                graph,
-                services.map { it.item },
-                config,
-                prefs
-            )
-            is JourneyCalculationTime.DepartureTime -> DepartureBasedRouter(
-                graph,
-                services.map { it.item },
-                config,
-                prefs
-            )
-        }
-    }
-
-}
